Fix datatable bugs
Deploy metadata did not set columns on first render which caused the data table to not render properly. This was fixed in the useDataTable to make sure that the columnOrder was reset properly and also fixed in the metadata data table to ensure the columns were available on the first render.

Updated rowHeight functions for regular data tables to match changes with latest version of react data table (only TreeGrid has a type property)

Increased column width on the automation control component

Fix incorrect use of isMatch with dateFns when loading time fields