Add Tab Visibility to Permissions Editor
Added tab visibility to permissions editor because this is often desired to be adjusted at the same time as other permissions

Some objects do not have a tab, in which case the user is unable to adjust tab permissions for those objects.

Tab visibility objects for standard objects have a prefix of 'standard-' instead of the full object name, in addition, the records need to be manually deleted to indicate no permissions.

resolves #663