Load: Show warning or error if record id is mapped #679
Give user a nice indicator that the load will most likely fail. We could have made this slightly better by not auto-mapping the record id for upsert, but there could be some edge-cases so the easy path was taken to show an error.

Mapping errors are considered "soft" as the uer can still continue if they so choose to.

resolves #679