Add Jetstream Organizations
Organizations allow users to group SFDC orgs so that they can work on groups of orgs at one time in isolation.

This feature is most useful for consultants working across many separate groups of orgs, but is also useful to separate out production from other sandboxes etc..

With our new authentication provider, it appears like we may not be able to support multiple accounts with the same email address which was the catalyst for this feature.

resolves #1018