Tighten up password reset process
Link userId to password reset token via database

Don't allow reset if the user does not have a password set - this is to protect against two users with duplicate email addresses

Don't allow a user to set up a password to login if another user with the same email has a password - that would lead to ambiguous login experience - this is just to ease the migration process to allow duplicate email addresses to be added in. Eventually we would ideally block this.

Moved google/sfdc sign up above other fields during login/sign up so we can nudge the user to do that instead of a password