Inline editing issues
Inline editing while filters were applied would cause edited values to be removed from the table and would remain hidden even after resetting all filters.

This was caused because the SET filter did not update to include the new value, so the record would be filtered out from the table - and even after resetting, the filter values were not re-calculated.

In addition the onRecordChanged emits the filtered values, not all values, so the parent component was saving just the filtered records thinking that was the universe - this also caused some (most?) records to not be available to be saved when submitting to Salesforce.

To solve, the data table component intercepts the change and ensures that the table set filters are updated to include the new value in the list and select the new record column value.

Unfortunately this only solves the set filter and does not solve any other filters (text, date, etc..) - that would require a complete refactor of how state is managed for the table, which is high-risk and a lot of work - so we will deal with it for now. To help this, a number has been added to the submit button to show the number of changed records.

resolves #1113