Refactor web extension to be Firefox compatible
Added polyfill to be able to use browser and chrome namespace interchangeably with promises

Refactored all data fetching to happen on the main page instead of intercepting the fetch event in a service worker, since that is not supported by Firefox and likely never will be supported

Refactored the authentication flow to use a content script since firefox does not support messaging from a web page directly to an extension.

Temporarily disabled support for downloading attachments since that CAN NEVER work in firefox with our service worker implementation

Added an axios adapter to handle fetching in a way that is compatible with the chrome extension router

Attempted to replace all instances of "chrome extension" with "browser extension" around the app, including in variable names.