Minify extension in dev mode to work around chrome issues
Chrome was reporting that a non-utf-8 character is included in the document, likely pulled in from the XLSX library since that is not part of the content script?

Is works fine when the code is minified, so not sure what the exact issue is or how to fix it, so working around it for now even though it slows down the development process.