Ensure automation control works if user has non-english language #1280
There was a query using the LIKE operator which apparently is matched against picklist labels, not values - and the labels are translated which means that the matching doesn't work as expected.

We are now hard-coding the TriggerType values instead of using LIKE in the query to ensure results are not language dependent.

resolves #1280