View bulk results does not work when bulk editing records from query results #1328
Remove nested modals for query download to ensure that view/download modal are rendered in the dom.

The parent modal was hidden and nothing in the dom was rendered, thus making the nested modals which were children to not be rendered.

Also ensured that the google file picker does not hide the parent modal just to ensure that there are no weird state issues with keeping track of the currently opened modal.

resolves #1328