Google Picker, Date Picker, Record Modal, Add to Changeset
Allow refreshing from google sheet on load records page even after user goes forward and backwards

Ensure date picker closes on outside click - there is still some slightly buggy behavior in a couple places. Inside a modal, the date picker closes when clicking on the date picker popup if the element is not focusable, and on the data table the date picker closes but not the entire edit input if user clicks outside of the table.

Refactored the ViewEditCreate record modal - this is now managed at the application level instead of managed at the component level - events are emitted from anywhere and the modal "JustWorks(TM)"

Any time a user views a record, the record search history now gets updated as a convenience for users to find their recently interacted with records

When uploading metadata to a changeset, we now allow the user to change some options and we default to rolling back on failure in production orgs (it is rare, but some users create outbound changesets in production orgs)

resolves #1344
resolves #1338
resolves #1340