Update application state pipeline
The desktop app was defaulting to localhost for the websocket server if history sync was enabled since it was falling back to the default cookie value.

The application cookie has been largely replaced by the heartbeat API, it still exists for web app, but can be removed in the future.

App state has been improved and centralized to be used in all form-factors instead of having a bunch of bespoke/duplicate implementations