Route validation for salesforce routes
Added validation to Salesforce domains when adding an org to avoid potential bad actors. Ensure that redirect is to valid Salesforce controlled domain.

Add additional check that org being accessed belongs to user. This is already checked at the DB level so the check is just an added redundancy.

Added hsts to force https for all domains and sub-domains.