Improve data fetching for related fields
Load related field metadata on-demand only when the object is selected in field mapping instead of pre-fetching all fields. This was done because because polymorphic fields have a lot of related objects and this caused pre-fetch of tons of data that was unnecessary.

When auto-mapping fields, data is pre-fetched for relationship fields as needed based on the headers in the CSV.