Improve desktop auto-update experience
The prior update experience was intrusive and annoying as it popped up a dialog that the user was forced to interact with. There were some cases where the dialog appeared multiple times and it was a bad experience.

Update status has been moved into a popover in the navbar to manage all states of the update process.