Improve binary attachment download process #1416
Remove reliance on service worker for binary downloads because it often required a browser refresh if the service worker was not registered correctly.

The file downloads and zip creation are now handed on the server using streaming with back-pressure control and timeout handling.

The desktop application uses a background job and the download happens on the main process where the file is saved directly to disk - we have a background job with really slick notification about the download progress.

The web extension is similar to the browser except the entire zip file is created and then saved since that is the only viable option.

In addition, support for StaticResource was added as a download option AND the user no longer needs to worry about what fields are included in the query - we re-query all the records to get the fields we rely on so the user can just do their thing without us getting in the way.

resolves #1416