Organization Group Improvements
Updated org experience to show expiring/expired orgs and let user refresh the connection easily.

When reconnecting an org, we utilize the login_hint for OAuth and we also pre-populate the org connection details to make the process more streamlined.

Renamed Jetstream Organization to Organization Group in the app and in code - this was done to avoid name confusion with a "salesforce organization" and a "(jetstream) organization". The app/docs/code variable names were all updated