Add links for installing connected app
There is a Salesforce bug where connected apps may not show up in some orgs (seems isolated to sandboxes) - added links to the page the user would have gone to if the app showed up in the oauth usage to our docs.

Also added a link to our docs from the "add org" popover just to add a really easy entry point for users to find those docs.

#1419